Move form editor routes to `/library/{slug}/editor`
