+++ designer/server/src/models/forms/library.js
@@ -1,3 +1,4 @@
+import config from '~/src/config.js'
 import * as forms from '~/src/lib/forms.js'
 
 export async function listViewModel() {
@@ -6,3 +7,24 @@ export async function listViewModel() {
 
   return { pageTitle, formItems }
 }
+
+/**
+ * @param {FormMetadata} metadata
+ */
+export function editorViewModel(metadata) {
+  const pageTitle = metadata.title
+
+  return {
+    backLink: {
+      text: 'Back to form library',
+      href: '/library'
+    },
+    pageTitle,
+    form: metadata,
+    previewUrl: config.previewUrl
+  }
+}
+
+/**
+ * @typedef {import('~/src/lib/forms.js').FormMetadata} FormMetadata
+ */
