+++ designer/server/src/routes/forms/index.js
@@ -1,6 +1,5 @@
 import api from './api.js'
 import create from './create.js'
-import editor from './editor.js'
 import library from './library.js'
 
-export default [api, create, editor, library].flat()
+export default [api, create, library].flat()
