+++ designer/server/src/views/forms/library.njk
@@ -29,7 +29,7 @@
 
     {% for form in formItems %}
       {% set formName %}
-        <a href="/editor/{{ form.slug | urlencode }}" class="govuk-link govuk-link--no-visited-state">
+        <a href="/library/{{ form.slug | urlencode }}/editor" class="govuk-link govuk-link--no-visited-state">
           {{ form.title }}
         </a>
       {% endset %}
